ml/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.15.0
numpy==1.26.4
Pillow==10.1.0
gunicorn==21.2.0 